New changes with color, content, and styles
Yellow was a very dull color, especially in header of the landing page.
I decided to give it a bit of fire and add a darker orange to it. As for
the other changes, I've cleaned up some comments and added more content
at the end to emphasize how the footer is always stuck to the bottom.
Note: This will be the first css project in my mini portfolio.